comparison between signed and unsigned integers

+++ moveit_ros/visualization/rviz_plugin_render_tools/src/trajectory_visualization.cpp
@@ -384,7 +384,7 @@ void TrajectoryVisualization::update(float wall_dt, float ros_dt)
       }
       else if (trajectory_slider_panel_ && trajectory_slider_panel_->isVisible())
       {
-        if (trajectory_slider_panel_->getSliderPosition() >= displaying_trajectory_message_->getWayPointCount() - 1)
+        if (static_cast<unsigned int>(trajectory_slider_panel_->getSliderPosition()) >= displaying_trajectory_message_->getWayPointCount() - 1)
           return;  // nothing more to do
         else
           animating_path_ = true;
